requirements.txt
numpy>=1.14.3
biopython>=1.71
pandas>=0.23.0
matplotlib>=2.2.2
